requirements.txt
streamlit 
langchain 
langchain_gemini
selenium
beautifulsoup4
lxml 
html5lib
python-dotenv
google-generativeai